requirements.txt
openai
requests
python-dotenv
alchemy-sdk
langchain
